Mitigate locking on infinite body stream
Previously, when a malicious request with a Content-Length set to
a value longer than actual length, the application would await until
the (infinite) bodyText stream would be completed.
Now, we set a sensible timeout on body reading stream and http response.
Then, the connection to a load balancer can be maintained but the request
is not blocking the connection anymore.